requirements.txt
streamlit
biopython
pandas
reportlab